+++ benchmarks/tst1d_19_envelope_wake_reduced_dispersion.py
@@ -6,7 +6,7 @@ nx = 192
 Lx = nx * dx
 npatch_x = 32
 laser_fwhm = 20. 
-center_laser = Lx-2.*laser_fwhm # the temporal center here is the same as waist position, but in principle they can differ
+center_laser = Lx-2.*laser_fwhm 
 time_start_moving_window =  0.