+++ src/ElectroMagn/LaserEnvelope2D.cpp
@@ -72,7 +72,7 @@ LaserEnvelope2D::LaserEnvelope2D( LaserEnvelope *envelope, Patch *patch, Params 
 }
 
 
-void LaserEnvelope2D::initEnvelope( Patch *patch, ElectroMagn *EMfields )
+void LaserEnvelope2D::initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields )
 {
     cField2D *A2D          = static_cast<cField2D *>( A_ );
     cField2D *A02D         = static_cast<cField2D *>( A0_ );
@@ -145,6 +145,58 @@ void LaserEnvelope2D::initEnvelope( Patch *patch, ElectroMagn *EMfields )
     
 }
 
+void LaserEnvelope2D::injectEnvelopeFromXmin( Patch *patch, Params &params, double time_dual )
+{
+
+    cField2D *A2D                 = static_cast<cField2D *>( A_ );
+    cField2D *A02D                = static_cast<cField2D *>( A0_ );
+
+    double t                      = time_dual;          // x-ct     , t=0
+    double t_previous_timestep    = time_dual-timestep; // x-c(t-dt), t=0
+
+    vector<double> position( 1, 0 );
+    double pos1 = cell_length[1]*( ( double )( patch->getCellStartingGlobalIndex( 1 ) )+( A_->isDual( 1 )?-0.5:0. ) );
+    
+    // oversize
+    int oversize_                 = params.oversize[0];
+
+    bool inject_envelope_from_this_patch = ( patch->isBoundary(0) ) && (  patch->isXmin() );
+
+    // Impose the envelope value for x=0 at time t and t-dt 
+    if ( inject_envelope_from_this_patch ){
+        position[0] = pos1;
+        for( unsigned int j=0 ; j<A_->dims_[1] ; j++ ) { 
+            ( *A2D  )( oversize_-1, j )  = profile_->complexValueAt( position, t );
+            ( *A02D )( oversize_-1, j )  = profile_->complexValueAt( position, t_previous_timestep );
+            position[0] += cell_length[1];
+        }
+        
+        if (envelope_solver=="explicit_reduced_dispersion"){
+            // This solver needs another point on the x direction, for one timestep, as initial condition. 
+            // This value will be found by locally solving the envelope wave equation.
+          
+            // Adapting the approach by C. Benedetti described in 
+            // F. Massimo et al., PPCF 2025 https://doi.org/10.1088/1361-6587/addc97,
+            // the second order derivatives in time and longitudinal coordinate are neglected
+            // under the paraxial and slowly varying envelope approximation.
+          
+            // The resulting envelope equation, that will be discretized with first order derivatives, is 
+            // \nabla^2_\perp A + 2*i*(dA/dx+dA/dt) = 0.
+            
+            // from A^n_i and A^(n-1)_i, we find A^n_{i-1}
+            for( unsigned int j=1 ; j<A_->dims_[1]-1 ; j++ ) {
+                // // A^n_{i-1} = A^n_i
+                ( *A2D  )( oversize_-2, j )  = ( *A2D  )( oversize_-1, j );
+                // A^n_{i-1}+= dx/dt*(A^{n+1}_i-A^{n}_i), until here it is like an upwind scheme for the advection equation
+                ( *A2D  )( oversize_-2, j ) += cell_length[0]/timestep*( ( *A2D  )( oversize_-1, j   )-   ( *A02D)( oversize_-1, j ));
+                // A^n_{i-1}+=dl/(2i) * (\nabla^2_\perp A)|^n_i
+                ( *A2D  )( oversize_-2, j ) += (cell_length[0]/2./i1) *( ( *A2D  )( oversize_-1, j-1 )-2.*( *A2D )( oversize_-1, j )+( *A2D )( oversize_-1, j+1 ) )*one_ov_dy_sq; // y part
+            } // end j loop
+            
+        }  // end if envelope_solver=="explicit_reduced_dispersion"
+    }
+      
+}
 
 LaserEnvelope2D::~LaserEnvelope2D()
 {