+++ src/ElectroMagn/LaserEnvelopeAM.cpp
@@ -74,7 +74,7 @@ LaserEnvelopeAM::LaserEnvelopeAM( LaserEnvelope *envelope, Patch *patch, Params 
 }
 
 
-void LaserEnvelopeAM::initEnvelope( Patch *patch, ElectroMagn *EMfields )
+void LaserEnvelopeAM::initEnvelopeInsideTheWindow( Patch *patch, ElectroMagn *EMfields )
 {
     cField2D *A2Dcyl          = static_cast<cField2D *>( A_ );
     cField2D *A02Dcyl         = static_cast<cField2D *>( A0_ );
@@ -174,6 +174,90 @@ void LaserEnvelopeAM::initEnvelope( Patch *patch, ElectroMagn *EMfields )
     }
 }
 
+void LaserEnvelopeAM::injectEnvelopeFromXmin( Patch *patch, Params &params, double time_dual )
+{
+  
+    cField2D *A2D                 = static_cast<cField2D *>( A_ );
+    cField2D *A02D                = static_cast<cField2D *>( A0_ );
+
+    double t                      = time_dual;          // x-ct     , t=0
+    double t_previous_timestep    = time_dual-timestep; // x-c(t-dt), t=0
+
+    vector<double>  position( 1, 0 );
+    //position[0] = cell_length[0]*( ( double )( patch->getCellStartingGlobalIndex( 0 ) )+( A_->isDual( 0 )?-0.5:0. ) + 1 );
+    double pos1 = cell_length[1]*( ( double )( patch->getCellStartingGlobalIndex( 1 ) )+( A_->isDual( 1 )?-0.5:0. ) );
+
+    // oversize
+    int oversize_                 = params.oversize[0];
+
+
+    bool inject_envelope_from_this_patch = ( patch->isBoundary(0) ) && (  patch->isXmin() );
+    bool isYmin = patch->isYmin();
+    
+    // Impose the envelope value for x=0 at time t and t-dt 
+    if ( inject_envelope_from_this_patch ){
+        position[0] = pos1;
+        for( unsigned int j=0 ; j<A_->dims_[1] ; j++ ) { 
+            ( *A2D  )( oversize_-1, j )  = profile_->complexValueAt( position, t );
+            ( *A02D )( oversize_-1, j )  = profile_->complexValueAt( position, t_previous_timestep );
+            position[0] += cell_length[1];
+        }
+        
+        if (isYmin){ // axis BC
+            unsigned int j = 2;  // j_p=2 corresponds to r=0    
+            ( *A2D  )( oversize_-1, j-1 )  = ( *A2D  )( oversize_-1, j+1 );
+            ( *A02D )( oversize_-1, j-2 )  = ( *A02D )( oversize_-1, j+2 );  
+        }
+        
+        if (envelope_solver=="explicit_reduced_dispersion"){
+            // This solver needs another point on the x direction, for one timestep, as initial condition. 
+            // This value will be found by locally solving the envelope wave equation.
+          
+            // Adapting the approach by C. Benedetti described in 
+            // F. Massimo et al., PPCF 2025 https://doi.org/10.1088/1361-6587/addc97,
+            // the second order derivatives in time and longitudinal coordinate are neglected
+            // under the paraxial and slowly varying envelope approximation.
+          
+            // The resulting envelope equation, that will be discretized with first order derivatives, is 
+            // \nabla^2_\perp A + 2*i*(dA/dl+dA/dt) = 0.
+            
+            int  j_glob = ( static_cast<ElectroMagnAM *>( patch->EMfields ) )->j_glob_;
+            
+            // from A^n_i and A^(n-1)_i, we find A^n_{i-1}
+            for( unsigned int j=std::max(3*isYmin,1) ; j<A_->dims_[1]-1 ; j++ ) { 
+            
+                // A^n_{i-1} = A^n_i
+                ( *A2D  )( oversize_-2, j )  = ( *A2D  )( oversize_-1, j );
+                // A^n_{i-1}+= dl/dt*(A^{n+1}_i-A^{n}_i), until here it is like an upwind scheme for the advection equation
+                ( *A2D  )( oversize_-2, j ) += cell_length[0]/timestep*( ( *A2D  )( oversize_-1, j   )-   ( *A02D)( oversize_-1, j ));
+                // A^n_{i-1}+=dl/(2i) * (\nabla^2_\perp A)|^n_i
+                ( *A2D  )( oversize_-2, j ) += (cell_length[0]/2./i1) *( ( *A2D  )( oversize_-1, j-1 )-2.*( *A2D )( oversize_-1, j )+( *A2D )( oversize_-1, j+1 ) )*one_ov_dr_sq; // r part
+                ( *A2D  )( oversize_-2, j ) += (cell_length[0]/2./i1) *( ( *A2D  )( oversize_-1, j+1 )-   ( *A2D )( oversize_-1, j-1 ) ) * one_ov_2dr / ( ( double )( j_glob+j )*dr ); // r part 
+            
+            } // end j loop
+            
+            if (isYmin){ // axis BC
+                unsigned int j = 2;  // j_p=2 corresponds to r=0 
+            
+                // here the transverse laplacian has a different discretization,
+                // otherwise, the integration is the same   
+                // A^n_{i-1} = A^n_i
+                ( *A2D  )( oversize_-2, j   )  = ( *A2D  )( oversize_-1, j );
+                // A^n_{i-1}+=dl/(2i) * (\nabla^2_\perp A)|^n_i
+                ( *A2D  )( oversize_-2, j   ) += (cell_length[0]/2./i1)*4.*( ( *A2D )( oversize_-1, j+1 )-( *A2D  )( oversize_-1, j ) ) * one_ov_dr_sq; // r part
+                
+                // boundary condition reflection below the axis
+                ( *A2D  )( oversize_-2, j-1 )  = ( *A2D  )( oversize_-2, j+1 );
+            
+            } // end axis BC
+            
+        }  // end if envelope_solver=="explicit_reduced_dispersion"
+          
+    } // end inject_envelope_from_this_patch
+    
+    
+      
+}
 
 LaserEnvelopeAM::~LaserEnvelopeAM()
 {