+++ src/ElectroMagnBC/EnvelopeBC3D_refl.cpp
@@ -55,7 +55,7 @@ void EnvelopeBC3D_refl::apply( LaserEnvelope *envelope, ElectroMagn *, Patch *pa
         for( unsigned int i=oversize_; i>0; i-- ) {
             for( unsigned int j=0 ; j<ny_p ; j++ ) {
                 for( unsigned int k=0 ; k<nz_p ; k++ ) {
-                    ( *A3D )( i-1, j, k ) = 0. ; // (*A3D)(i,j,k);
+                    ( *A3D )  ( i-1, j, k ) = 0. ; // (*A3D)(i,j,k);
                     ( *Phi3D )( i-1, j, k ) = 0. ; // std::abs((*A3D)  (i,j,k)) * std::abs((*A3D)  (i,j,k)) * 0.5;
                 }//k
             }//j
@@ -68,7 +68,7 @@ void EnvelopeBC3D_refl::apply( LaserEnvelope *envelope, ElectroMagn *, Patch *pa
         for( unsigned int i=nx_p-oversize_; i<nx_p; i++ ) {
             for( unsigned int j=0 ; j<ny_p ; j++ ) {
                 for( unsigned int k=0 ; k<nz_p ; k++ ) {
-                    ( *A3D )( i, j, k ) = 0. ; //(*A3D)(i-1,j,k);
+                    ( *A3D )  ( i, j, k ) = 0. ; //(*A3D)(i-1,j,k);
                     ( *Phi3D )( i, j, k ) = 0. ; //std::abs((*A3D)  (i-1,j,k)) * std::abs((*A3D)  (i-1,j,k)) * 0.5;
                 }//k
             }//j
@@ -81,7 +81,7 @@ void EnvelopeBC3D_refl::apply( LaserEnvelope *envelope, ElectroMagn *, Patch *pa
         for( unsigned int i=0; i<nx_p; i++ ) {
             for( unsigned int j=oversize_ ; j>0 ; j-- ) {
                 for( unsigned int k=0; k<nz_p; k++ ) {
-                    ( *A3D )( i, j-1, k ) = 0. ; // (*A3D)(i,j,k);
+                    ( *A3D )  ( i, j-1, k ) = 0. ; // (*A3D)(i,j,k);
                     ( *Phi3D )( i, j-1, k ) = 0. ; // std::abs((*A3D)  (i,j,k)) * std::abs((*A3D)  (i,j,k)) * 0.5;
                 }//k
             }//j
@@ -94,7 +94,7 @@ void EnvelopeBC3D_refl::apply( LaserEnvelope *envelope, ElectroMagn *, Patch *pa
         for( unsigned int i=0; i<nx_p; i++ ) {
             for( unsigned int j=ny_p-oversize_; j<ny_p ; j++ ) {
                 for( unsigned int k=0; k<nz_p; k++ ) {
-                    ( *A3D )( i, j, k ) = 0. ; // (*A3D)(i,j-1,k);
+                    ( *A3D )  ( i, j, k ) = 0. ; // (*A3D)(i,j-1,k);
                     ( *Phi3D )( i, j, k ) = 0. ; // std::abs((*A3D)  (i,j-1,k)) * std::abs((*A3D)  (i,j-1,k)) * 0.5;
                 }//k
             }//j
@@ -107,7 +107,7 @@ void EnvelopeBC3D_refl::apply( LaserEnvelope *envelope, ElectroMagn *, Patch *pa
         for( unsigned int i=0; i<nx_p; i++ ) {
             for( unsigned int j=0; j<ny_p; j++ ) {
                 for( unsigned int k=oversize_; k>0; k-- ) {
-                    ( *A3D )( i, j, k-1 ) = 0. ; // (*A3D)(i,j,k);
+                    ( *A3D   )( i, j, k-1 ) = 0. ; // (*A3D)(i,j,k);
                     ( *Phi3D )( i, j, k-1 ) = 0. ; // std::abs((*A3D)  (i,j,k)) * std::abs((*A3D)  (i,j,k)) * 0.5;
                 }//k
             }//j
@@ -120,7 +120,7 @@ void EnvelopeBC3D_refl::apply( LaserEnvelope *envelope, ElectroMagn *, Patch *pa
         for( unsigned int i=0; i<nx_p; i++ ) {
             for( unsigned int j=0; j<ny_p; j++ ) {
                 for( unsigned int k=nz_p-oversize_; k<nz_p; k++ ) {
-                    ( *A3D )( i, j, k ) = 0. ; // (*A3D)(i,j,k-1);
+                    ( *A3D   )( i, j, k ) = 0. ; // (*A3D)(i,j,k-1);
                     ( *Phi3D )( i, j, k ) = 0. ; // std::abs((*A3D)  (i,j,k-1)) * std::abs((*A3D)  (i,j,k-1)) * 0.5;
                 }//z
             }//j