+++ src/ElectroMagnSolver/MF_SolverAM_Terzani.cpp
@@ -52,11 +52,12 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         bool isXmin  = ( static_cast<ElectroMagnAM *>( fields ) )->isXmin;
         bool isXmax  = ( static_cast<ElectroMagnAM *>( fields ) )->isXmax;
         bool isYmin  = ( static_cast<ElectroMagnAM *>( fields ) )->isYmin;
+        int oversize_r = fields->oversize[1];
 
         // ---- Magnetic field Bl^(p,d)
         for( unsigned int i=0 ; i<nl_p;  i++ ) {
             #pragma omp simd
-            for( unsigned int j=1+isYmin*2 ; j<nr_d-1 ; j++ ) {
+            for( unsigned int j=1+isYmin*oversize_r ; j<nr_d-1 ; j++ ) {
                 ( *Bl )( i, j ) += - dt/( ( j_glob+j-0.5 )*dr ) * ( ( static_cast<double>( j+j_glob ))*( *Et )( i, j ) 
                                    - ( static_cast<double>(j+j_glob-1. ))*( *Et )( i, j-1 ) 
                                    + Icpx*(static_cast<double>(imode))   *( *Er )( i, j ) );
@@ -66,7 +67,7 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         // ---- Magnetic field Br^(d,p), smaller loop boundaries to avoid segfault
         for( unsigned int i=2 ; i<nl_d-2 ; i++ ) {
             #pragma omp simd
-            for( unsigned int j=isYmin*3 ; j<nr_p ; j++ ) { //Specific condition on axis
+            for( unsigned int j=isYmin*(oversize_r+1) ; j<nr_p ; j++ ) { //Specific condition on axis
                 ( *Br )( i, j ) += dt_ov_dl * (1.-3.*delta) * ( ( *Et )( i  , j ) - ( *Et )( i-1, j ) )
                                  + dt_ov_dl * delta         * ( ( *Et )( i+1, j ) - ( *Et )( i-2, j ) )
                                 +Icpx*dt*( static_cast<double>(imode) )/( ( static_cast<double>( j_glob+j ))*dr )*( *El )( i, j ) ;
@@ -75,7 +76,7 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         // Magnetic field Br^(d,p), left border of the patch: evolve as in Yee solver
         for( unsigned int i=1 ; i<2 ; i++ ) {
             #pragma omp simd 
-            for( unsigned int j=isYmin*3 ; j<nr_p ; j++ ) { //Specific condition on axis
+            for( unsigned int j=isYmin*(oversize_r+1) ; j<nr_p ; j++ ) { //Specific condition on axis
                 ( *Br )( i, j ) += dt_ov_dl * ( ( *Et )( i, j ) - ( *Et )( i-1, j ) )
                                 +Icpx*dt*( static_cast<double>(imode) )/( ( static_cast<double>( j_glob+j )) *dr )*( *El )( i, j ) ;
             }
@@ -83,7 +84,7 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         // Magnetic field Br^(d,p), right border of the patch: evolve as in Yee solver
         for( unsigned int i=nl_d-2 ; i<nl_d-1 ; i++ ) {
             #pragma omp simd 
-            for( unsigned int j=isYmin*3 ; j<nr_p ; j++ ) { //Specific condition on axis
+            for( unsigned int j=isYmin*(oversize_r+1) ; j<nr_p ; j++ ) { //Specific condition on axis
                 ( *Br )( i, j ) += dt_ov_dl * ( ( *Et )( i, j ) - ( *Et )( i-1, j ) )
                                  +Icpx*dt*( static_cast<double>(imode) )/( ( static_cast<double>( j_glob+j )) *dr )*( *El )( i, j ) ;
             }
@@ -92,7 +93,7 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         // ---- Magnetic field Bt^(d,d), smaller loop boundaries to avoid segfault
         for( unsigned int i=2 ; i<nl_d-2 ; i++ ) {
             #pragma omp simd
-            for( unsigned int j=1 + isYmin*2 ; j<nr_d-1 ; j++ ) {
+            for( unsigned int j=1 + isYmin*oversize_r ; j<nr_d-1 ; j++ ) {
                 ( *Bt )( i, j ) += dt_ov_dr *                ( ( *El )( i, j   ) - ( *El )( i, j-1 ) )
                                   -dt_ov_dl * (1.-3*delta) * ( ( *Er )( i, j   ) - ( *Er )( i-1, j ) )
                                   -dt_ov_dl * delta        * ( ( *Er )( i+1, j ) - ( *Er )( i-2, j ) );
@@ -102,7 +103,7 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         // Magnetic field Bt^(d,d), left border of the patch : evolve as in Yee solver
         for( unsigned int i=1 ; i<2 ; i++ ) {
             #pragma omp simd 
-            for( unsigned int j=1 + isYmin*2 ; j<nr_d-1 ; j++ ) {
+            for( unsigned int j=1 + isYmin*oversize_r ; j<nr_d-1 ; j++ ) {
                 ( *Bt )( i, j ) += dt_ov_dr * ( ( *El )( i, j ) - ( *El )( i, j-1 ) )
                                   -dt_ov_dl * ( ( *Er )( i, j ) - ( *Er )( i-1, j ) );
             }
@@ -111,7 +112,7 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         // Magnetic field Bt^(d,d), right border of the patch : evolve as in Yee solver
         for( unsigned int i=nl_d-2 ; i<nl_d-1 ; i++ ) {
           #pragma omp simd 
-          for( unsigned int j=1 + isYmin*2 ; j<nr_d-1 ; j++ ) {
+          for( unsigned int j=1 + isYmin*oversize_r ; j<nr_d-1 ; j++ ) {
               ( *Bt )( i, j ) +=  dt_ov_dr * ( ( *El )( i, j ) - ( *El )( i, j-1 ) )
                                  -dt_ov_dl * ( ( *Er )( i, j ) - ( *Er )( i-1, j ) );
           }
@@ -121,69 +122,68 @@ void MF_SolverAM_Terzani::operator()( ElectroMagn *fields )
         // On axis conditions: as in Yee solver, 
         // except for mode 1 where derivative along x is present
         if( isYmin ) {
-            unsigned int j=2;
             if( imode==0 ) {
                 // ---- Br
                 for( unsigned int i=0 ; i<nl_d ; i++ ) {
-                    ( *Br )( i, j )= 0;
+                    ( *Br )( i, oversize_r )= 0;
                     ( *Br )( i, 1 )=-( *Br )( i, 3 );
                 }
                 // ---- Bt
                 for( unsigned int i= 0 ; i<nl_d ; i++ ) {
-                    ( *Bt )( i, j )= -( *Bt )( i, j+1 );
+                    ( *Bt )( i, oversize_r )= -( *Bt )( i, oversize_r+1 );
                 }
                 // ---- Bl
                 for( unsigned int i= 0 ; i<nl_p ; i++ ) {
-                    ( *Bl )( i, j )= ( *Bl )( i, j+1 );
+                    ( *Bl )( i, oversize_r )= ( *Bl )( i, oversize_r+1 );
                 }
             }
 
             else if( imode==1 ) {
                 // ---- Bl
                 for( unsigned int i=0 ; i<nl_p  ; i++ ) {
-                    ( *Bl )( i, j )= -( *Bl )( i, j+1 ); // Zero Bl mode 1 on axis.
+                    ( *Bl )( i, oversize_r )= -( *Bl )( i, oversize_r+1 ); // Zero Bl mode 1 on axis.
                 }
                 
                 // ---- Br: use Terzani's derivative in smaller loop boundaries
                 for( unsigned int i=2 ; i<nl_d-2 ; i++ ) {
-                    ( *Br )( i, j )+=  Icpx*dt_ov_dr*( *El )( i, j+1 )
-                                   +	 (1.-3.*delta)*	dt_ov_dl * ( ( *Et )( i  , j )-( *Et )( i-1, j ) )
-                                   +   delta        * dt_ov_dl * ( ( *Et )( i+1, j )-( *Et )( i-2, j ) );
+                    ( *Br )( i, oversize_r )+=  Icpx*dt_ov_dr*( *El )( i, oversize_r+1 )
+                                   +	 (1.-3.*delta)*	dt_ov_dl * ( ( *Et )( i  , oversize_r )-( *Et )( i-1, oversize_r ) )
+                                   +   delta        * dt_ov_dl * ( ( *Et )( i+1, oversize_r )-( *Et )( i-2, oversize_r ) );
                     ( *Br )( i, 1 ) =( *Br )( i, 3 );
                 }
                 
                 // ---- Br: as in Yee solver at the left border of the patch 
                 for( unsigned int i=1 ; i<2 ; i++ ) {
-                    ( *Br )( i, j )+=  Icpx*dt_ov_dr*( *El )( i, j+1 )
-                                       +	dt_ov_dl*( ( *Et )( i, j   )-( *Et )( i-1, j ) );
+                    ( *Br )( i, oversize_r )+=  Icpx*dt_ov_dr*( *El )( i, oversize_r+1 )
+                                       +	dt_ov_dl*( ( *Et )( i, oversize_r   )-( *Et )( i-1, oversize_r ) );
                     ( *Br )( i, 1 ) =( *Br )( i, 3 );
                 }
                 
                 // ---- Br: as in Yee solver at the right border of the patch 
                 for( unsigned int i=nl_d-2 ; i<nl_d-1 ; i++ ) {
-                    ( *Br )( i, j )+=  Icpx*dt_ov_dr*( *El )( i, j+1 )
-                                       +	dt_ov_dl*( ( *Et )( i, j   )-( *Et )( i-1, j ) );
+                    ( *Br )( i, oversize_r )+=  Icpx*dt_ov_dr*( *El )( i, oversize_r+1 )
+                                       +	dt_ov_dl*( ( *Et )( i, oversize_r   )-( *Et )( i-1, oversize_r ) );
                     ( *Br )( i, 1 ) =( *Br )( i, 3 );
                 }
                 
                 // ---- Bt
                 for( unsigned int i=0; i<nl_d ; i++ ) {
-                    ( *Bt )( i, j )= ( *Bt )( i, j+1 ); // Non zero Bt mode 1 on axis.
+                    ( *Bt )( i, oversize_r )= ( *Bt )( i, oversize_r+1 ); // Non zero Bt mode 1 on axis.
                 }
 
             } else { // modes > 1
                 // ---- Bl
                 for( unsigned int  i=0 ; i<nl_p; i++ ) {
-                    ( *Bl )( i, j )= -( *Bl )( i, j+1 );
+                    ( *Bl )( i, oversize_r )= -( *Bl )( i, oversize_r+1 );
                 }
                 // ---- Br
                 for( unsigned int i=0 ; i<nl_d; i++ ) {
-                    ( *Br )( i, j )= 0;
+                    ( *Br )( i, oversize_r )= 0;
                     ( *Br )( i, 1 )=-( *Br )( i, 3 );
                 }
                 // ---- Bt
                 for( unsigned int  i=0 ; i<nl_d ; i++ ) {
-                    ( *Bt )( i, j )= - ( *Bt )( i, j+1 );
+                    ( *Bt )( i, oversize_r )= - ( *Bt )( i, oversize_r+1 );
                 }
             }
         }