+++ src/MovWindow/SimWindow.cpp
@@ -258,6 +258,7 @@ void SimWindow::shift( VectorPatch &vecPatches, SmileiMPI *smpi, Params &params,
                             }
                         }
                         mypatch->EMfields->envelope->EnvBoundCond = EnvelopeBC_Factory::create( params, mypatch );
+                        mypatch->EMfields->envelope->keep_injecting_laser_envelope = false;
                     }
 
                     mypatch->EMfields->laserDisabled();
@@ -319,6 +320,7 @@ void SimWindow::shift( VectorPatch &vecPatches, SmileiMPI *smpi, Params &params,
                             }
                         }
                         mypatch->EMfields->envelope->EnvBoundCond = EnvelopeBC_Factory::create( params, mypatch );
+                        mypatch->EMfields->envelope->keep_injecting_laser_envelope = false;
                     }
 
                     mypatch->EMfields->laserDisabled();
@@ -338,6 +340,7 @@ void SimWindow::shift( VectorPatch &vecPatches, SmileiMPI *smpi, Params &params,
                             }
                         }
                         mypatch->EMfields->envelope->EnvBoundCond = EnvelopeBC_Factory::create( params, mypatch );
+                        mypatch->EMfields->envelope->keep_injecting_laser_envelope = false;
                     }
 
                     mypatch->EMfields->laserDisabled();
@@ -422,6 +425,7 @@ void SimWindow::shift( VectorPatch &vecPatches, SmileiMPI *smpi, Params &params,
 #endif
     for( unsigned int ipatch = 0 ; ipatch < nPatches ; ipatch++ ) {
             mypatch = vecPatches.patches_[ipatch];
+            if (mypatch->EMfields->envelope) mypatch->EMfields->envelope->keep_injecting_laser_envelope = false;
             if( mypatch->isXmin() ) {
                 if (!params.multiple_decomposition){
                     mypatch->EMfields->emBoundCond[0]->apply(mypatch->EMfields, time_dual, mypatch);